feat: SchemaMigrationPlan で Task.order を Int にマイグレーション
- SchemaV1: order なしの Task, TaskStep
- SchemaV2: order: Int を追加した Task, TaskStep
- SchemaLatest typealias でアプリ全体から参照
- BabyStepsMigrationPlan: didMigrate で createdAt 順に order を付与
