+++ Sources/Models/MigrationPlan.swift
@@ -1,0 +1,28 @@
+import Foundation
+import SwiftData
+
+enum BabyStepsMigrationPlan: SchemaMigrationPlan {
+    static var schemas: [any VersionedSchema.Type] {
+        [SchemaV1.self, SchemaV2.self]
+    }
+
+    static var stages: [MigrationStage] {
+        [migrateV1toV2]
+    }
+
+    static let migrateV1toV2 = MigrationStage.custom(
+        fromVersion: SchemaV1.self,
+        toVersion: SchemaV2.self,
+        willMigrate: { _ in },
+        didMigrate: { context in
+            let descriptor = FetchDescriptor<SchemaV2.Task>(
+                sortBy: [SortDescriptor(\.createdAt, order: .forward)]
+            )
+            let tasks = try context.fetch(descriptor)
+            for (index, task) in tasks.enumerated() {
+                task.order = index
+            }
+            try context.save()
+        }
+    )
+}
