+++ Sources/Models/SchemaLatest.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+typealias SchemaLatest = SchemaV2
+
+// MARK: - Models
+
+typealias Task = SchemaLatest.Task
+typealias TaskStep = SchemaLatest.TaskStep
